requirements.txt
absl-py==0.11.0
aiohttp @ file:///C:/ci/aiohttp_1614361133907/work
alpaca-trade-api==0.51.0
appdirs==1.4.4
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
beautifulsoup4==4.9.3
blinker==1.4
bokeh @ file:///C:/ci/bokeh_1603297932359/work
brotlipy==0.7.0
cachetools==4.2.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1613247308275/work
chardet==4.0.0
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
coverage @ file:///C:/ci/coverage_1614614910274/work
cryptography @ file:///C:/ci/cryptography_1616769432139/work
cycler==0.10.0
Cython @ file:///C:/ci/cython_1618435342622/work
distlib==0.3.1
Django @ file:///tmp/build/80754af9/django_1606860386887/work
django-crispy-forms==1.11.0
filelock==3.0.12
flatbuffers==1.12
gast==0.3.3
google==3.0.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata==3.3.0
Jinja2==2.11.2
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
Keras @ file:///tmp/build/80754af9/keras_split_1593112142734/work
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1612282618948/work
lxml==4.6.2
Markdown==3.3.3
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib @ file:///C:/ci/matplotlib-suite_1613408055530/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict @ file:///C:/ci/multidict_1607367865052/work
multitasking==0.0.9
newsapi-python==0.2.6
numpy==1.19.3
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1607971725249/work
pandas @ file:///C:/ci/pandas_1609008089615/work
pandas-datareader==0.9.0
patsy==0.5.1
Pillow @ file:///C:/ci/pillow_1609786872067/work
protobuf==3.14.0
psycopg2 @ file:///C:/ci/psycopg2_1608129576564/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyChart.JS==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyreadline==2.1
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
PyYAML==5.3.1
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn @ file:///C:/ci/scikit-learn_1614446716349/work
scipy @ file:///C:/ci/scipy_1616703425152/work
sequential==1.0.0
six @ file:///C:/ci/six_1605205426665/work
soupsieve==2.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
statsmodels==0.12.2
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_te/tensorboard_1614593728657/work/tmp_pip_dir
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator @ file:///tmp/build/80754af9/tensorflow-estimator_1599136169057/work/whl_temp/tensorflow_estimator-2.3.0-py2.py3-none-any.whl
termcolor==1.1.0
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tornado @ file:///C:/ci/tornado_1606935947090/work
typing-extensions==3.7.4.3
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
virtualenv==20.2.2
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt==1.12.1
yarl @ file:///C:/ci/yarl_1606940114600/work
yfinance==0.1.55
zipp==3.4.0
